Some programs needed for native gcc bootstrap
    For expr program:
      Adding asprintf and dependencies to stdio

    General:
      Fixed int64_t and uint64_t types
      Fixed regoff_t type (it's a 32 bit integer)

    Note that expr and uname, used by configure scripts,
    are not in default APEX building. This is part of a future
    big change in what will be involved a havery.inc file
    for native building, arch header splitting and more.

    expr and uname will be into build essentials package
    as binaries like the rest for now. Keeping their source code
    here in repo.

+++ include/regex.h
@@ -15,7 +15,7 @@ extern "C" {
 
 #include <features.h>
 
-typedef size_t regoff_t;
+typedef int regoff_t;
 
 typedef struct re_pattern_buffer {
 	size_t re_nsub;
