+++ include/stdint.h
@@ -18,8 +18,8 @@ typedef	unsigned short	uint16_t;
 typedef	signed short	int16_t;
 typedef unsigned int	uint32_t;
 typedef signed int	int32_t;
-typedef	unsigned long uint64_t;
-typedef	long	int64_t;
+typedef	unsigned long long uint64_t;
+typedef	long long	int64_t;
 typedef uint32_t	usize;
 
 typedef int8_t int_fast8_t;
